requirements.txt
# 核心依赖
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.0

# 数据库
asyncpg==0.29.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 工具库
aiohttp==3.9.1
pillow==10.1.0
numpy==1.24.3
pandas==2.1.4

# 开发工具
python-dotenv==1.0.0